+++ quantforge/__init__.py
@@ -726,6 +726,7 @@ from .bessel2 import (
 from .reverse_ad import Var, reverse_gradient
 from .lbfgs import lbfgs
 from .reverse_jacobian import reverse_jacobian, reverse_hessian, reverse_gradient_vector
+from .newton_min import newton_min
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1430,7 +1431,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.12.0"
+__version__ = "5.13.0"
 
 __all__ = [
     "OptionType",
@@ -2428,6 +2429,7 @@ __all__ = [
     "reverse_hessian",
     "reverse_gradient_vector",
     "lbfgs",
+    "newton_min",
     "Dual",
     "dual_derivative",
     "dual_gradient",